:recycle: refactor(test): split monolithic test files into per-language files
JVM tests (8 files, 180 tests):
  RomanizationJapaneseTest  — 23 tests
  RomanizationChineseTest   — 4 tests
  RomanizationKoreanTest    — 38 tests
  RomanizationHindiTest     — 25 tests
  RomanizationPunjabiTest   — 13 tests
  RomanizationCyrillicTest  — 39 tests
  RomanizationDetectionTest — 25 tests
  RomanizationAutoDetectTest— 13 tests

Android tests (8 files, 128 tests): same structure,
each with @BeforeClass dictionary loading via companion object.
