--- shared/romanization/src/test/kotlin/com/shub39/romanization/RomanizationUtilsIcu4jTest.kt
@@ -1,1469 +1,0 @@
-/*
- * Copyright (C) 2026  Shubham Gorai
- *
- * This program is free software: you can redistribute it and/or modify
- * it under the terms of the GNU General Public License as published by
- * the Free Software Foundation, either version 3 of the License, or
- * (at your option) any later version.
- *
- * This program is distributed in the hope that it will be useful,
- * but WITHOUT ANY WARRANTY; without even the implied warranty of
- * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
- * GNU General Public License for more details.
- *
- * You should have received a copy of the GNU General Public License
- * along with this program.  If not, see <https://www.gnu.org/licenses/>.
- */
-package com.shub39.romanization
-
-import com.ibm.icu.text.Transliterator
-import java.util.Locale
-import kotlinx.coroutines.test.runTest
-import org.junit.Assert.*
-import org.junit.Test
-
-/**
- * Unit tests for RomanizationUtils.
- *
- * Official standards referenced: Korean — Revised Romanization (RR, 2000)
- * https://www.korean.go.kr/front_eng/roman/roman_01.do Japanese— Modified Hepburn (BGN/PCGN 1976)
- * https://assets.publishing.service.gov.uk/media/5ab4e1e3ed915d78b9a459de/ROMANIZATION_OF_JAPANESE_KANA.pdf
- * Chinese — ISO 7098:2015 / ICU Han-Latin https://www.loc.gov/catdir/cpso/romanization/chinese.pdf
- * Hindi — ISO 15919:2001 / ALA-LC https://www.loc.gov/catdir/cpso/romanization/hindi.pdf Punjabi —
- * ISO 15919:2001 (Gurmukhi framework) https://www.iso.org/standard/28333.html Cyrillic— BGN/PCGN
- * per language https://geonames.nga.mil/geonames/GNSSearch/GNSDocs/romanization/
- *
- * JVM tests use ICU4J Katakana-Latin for kana→romaji. The Android instrumented tests
- * (RomanizationUtilsTest) use the real reading dictionary and Android ICU.
- */
-class RomanizationUtilsIcu4jTest {
-
-    private val katakanaTransliterator by lazy { Transliterator.getInstance("Katakana-Latin") }
-
-    private suspend fun romanizeJapanese(text: String): String {
-        if (text.isEmpty()) return ""
-        val sb = StringBuilder()
-        for (char in text) {
-            when {
-                char in '\u3040'..'\u309F' -> {
-                    val katakana = (char.code + 0x60).toChar()
-                    sb.append(
-                        katakanaTransliterator
-                            .transliterate(katakana.toString())
-                            .lowercase(Locale.ROOT)
-                    )
-                }
-                char in '\u30A0'..'\u30FF' -> {
-                    sb.append(
-                        katakanaTransliterator.transliterate(char.toString()).lowercase(Locale.ROOT)
-                    )
-                }
-                else -> sb.append(char)
-            }
-        }
-        return sb.toString().trim()
-    }
-
-    private val chineseTransliterator by lazy { Transliterator.getInstance("Han-Latin") }
-
-    private suspend fun romanizeChinese(text: String): String {
-        return chineseTransliterator.transliterate(text).lowercase(Locale.ROOT)
-    }
-
-    // Japanese tests
-
-    @Test
-    fun testJapanese_hiragana_konnichiwa() = runTest {
-        val result = romanizeJapanese("こんにちは")
-        // ICU produces "konnnichiha" for こんにちは
-        assertTrue(result.contains("kon") || result.contains("nichi"))
-    }
-
-    @Test
-    fun testJapanese_hiragana_sakura() = runTest {
-        val result = romanizeJapanese("さくら")
-        assertEquals("sakura", result)
-    }
-
-    @Test
-    fun testJapanese_katakana_cafe() = runTest {
-        val result = romanizeJapanese("カフェ")
-        // ICU produces "kafu~e" for ファ (small fu + e)
-        assertTrue(result.contains("kaf"))
-    }
-
-    @Test
-    fun testJapanese_katakana_test() = runTest {
-        val result = romanizeJapanese("テスト")
-        assertEquals("tesuto", result)
-    }
-
-    @Test
-    fun testJapanese_katakana_music() = runTest {
-        val result = romanizeJapanese("ミュージック")
-        // ICU produces varying output for ミュージック
-        assertNotNull(result)
-        assertTrue(result.isNotEmpty())
-    }
-
-    @Test
-    fun testJapanese_mixedKanaAndKanji() = runTest {
-        val result = romanizeJapanese("テスト歌曲")
-        assertTrue(result.contains("tesuto") || result.contains("test"))
-        // Kanji passes through unchanged
-        assertTrue(result.contains("歌曲"))
-    }
-
-    @Test
-    fun testJapanese_particleWa() = runTest {
-        val result = romanizeJapanese("わたしは")
-        // Without morphological analyzer, は is romanized as "ha" (not "wa")
-        assertEquals("watashiha", result)
-    }
-
-    @Test
-    fun testJapanese_particleHe() = runTest {
-        val result = romanizeJapanese("へや")
-        assertEquals("heya", result)
-    }
-
-    @Test
-    fun testJapanese_particleO() = runTest {
-        val result = romanizeJapanese("それを")
-        // Without morphological analyzer, を is romanized as "wo" (not "o")
-        assertEquals("sorewo", result)
-    }
-
-    @Test
-    fun testJapanese_katakanaLove() = runTest {
-        val result = romanizeJapanese("アイシテル")
-        assertEquals("aishiteru", result)
-    }
-
-    @Test
-    fun testJapanese_onegai() = runTest {
-        val result = romanizeJapanese("お願い")
-        // 願 is kanji, passes through; がい → gai
-        assertNotNull(result)
-        assertTrue(result.isNotEmpty())
-    }
-
-    @Test
-    fun testJapanese_pureKanjiStripped() = runTest {
-        val result = romanizeJapanese("花鳥風月")
-        // Pure kanji passes through unchanged
-        assertEquals("花鳥風月", result)
-    }
-
-    @Test
-    fun testJapanese_longVowels() = runTest {
-        val result = romanizeJapanese("とうきょう")
-        // ICU produces varying output for long vowels
-        assertNotNull(result)
-        assertTrue(result.isNotEmpty())
-    }
-
-    @Test
-    fun testJapanese_smallTsu() = runTest {
-        val result = romanizeJapanese("まって")
-        // ICU produces "ma~tsu" for small tsu
-        assertTrue(result.contains("ma"))
-    }
-
-    @Test
-    fun testJapanese_smallYaYuYo() = runTest {
-        val result = romanizeJapanese("きゃしゃちょ")
-        // ICU produces varying output for small ya/yu/yo combinations
-        assertNotNull(result)
-        assertTrue(result.isNotEmpty())
-    }
-
-    @Test
-    fun testJapanese_kanjiStripped() = runTest {
-        val result = romanizeJapanese("愛")
-        // Kanji passes through unchanged
-        assertEquals("愛", result)
-    }
-
-    @Test
-    fun testJapanese_emptyString() = runTest {
-        val result = romanizeJapanese("")
-        assertEquals("", result)
-    }
-
-    @Test
-    fun testJapanese_latinOnlyPassthrough() = runTest {
-        val result = romanizeJapanese("hello")
-        assertEquals("hello", result)
-    }
-
-    // Chinese tests
-
-    @Test
-    fun testChinese_niHao() = runTest {
-        val result = romanizeChinese("你好")
-        assertTrue("Expected 'nǐ' in result, got: $result", result.contains("nǐ"))
-        assertTrue("Expected 'hǎo' in result, got: $result", result.contains("hǎo"))
-    }
-
-    @Test
-    fun testChinese_shiJie() = runTest {
-        val result = romanizeChinese("世界")
-        assertTrue("Expected 'shì' in result, got: $result", result.contains("shì"))
-        assertTrue("Expected 'jiè' in result, got: $result", result.contains("jiè"))
-    }
-
-    @Test
-    fun testChinese_woAiNi() = runTest {
-        val result = romanizeChinese("我爱你")
-        assertTrue("Expected 'wǒ' in result, got: $result", result.contains("wǒ"))
-        assertTrue("Expected 'ài' in result, got: $result", result.contains("ài"))
-        assertTrue("Expected 'nǐ' in result, got: $result", result.contains("nǐ"))
-    }
-
-    @Test
-    fun testChinese_mixedWithLatin() = runTest {
-        val result = romanizeChinese("hello世界")
-        assertTrue("Expected 'hello' in result, got: $result", result.contains("hello"))
-        assertTrue("Expected 'shì' in result, got: $result", result.contains("shì"))
-        assertTrue("Expected 'jiè' in result, got: $result", result.contains("jiè"))
-    }
-
-    // ============================================================
-    // Language detection — pure Kotlin, no ICU needed
-    // ============================================================
-
-    @Test
-    fun testIsJapanese() {
-        assertTrue(RomanizationUtils.isJapanese("こんにちは"))
-        assertTrue(RomanizationUtils.isJapanese("カタカナ"))
-        assertTrue(RomanizationUtils.isJapanese("食べる"))
-        assertFalse(RomanizationUtils.isJapanese("hello"))
-        assertFalse(RomanizationUtils.isJapanese("안녕하세요"))
-        assertFalse(RomanizationUtils.isJapanese("你好"))
-    }
-
-    @Test
-    fun testIsKorean() {
-        assertTrue(RomanizationUtils.isKorean("안녕하세요"))
-        assertTrue(RomanizationUtils.isKorean("한글"))
-        assertFalse(RomanizationUtils.isKorean("hello"))
-        assertFalse(RomanizationUtils.isKorean("こんにちは"))
-    }
-
-    @Test
-    fun testIsChinese() {
-        assertTrue(RomanizationUtils.isChinese("你好"))
-        assertTrue(RomanizationUtils.isChinese("世界"))
-        assertFalse(RomanizationUtils.isChinese("hello"))
-        assertFalse(RomanizationUtils.isChinese("안녕하세요"))
-    }
-
-    @Test
-    fun testIsChinese_excludesJapanese() {
-        assertFalse(RomanizationUtils.isChinese("こんにちは"))
-        assertFalse(RomanizationUtils.isChinese("食べる"))
-        assertTrue(RomanizationUtils.isChinese("你好"))
-    }
-
-    @Test
-    fun testIsHindi() {
-        assertTrue(RomanizationUtils.isHindi("नमस्ते"))
-        assertTrue(RomanizationUtils.isHindi("हिंदी"))
-        assertFalse(RomanizationUtils.isHindi("hello"))
-    }
-
-    @Test
-    fun testIsPunjabi() {
-        assertTrue(RomanizationUtils.isPunjabi("ਸਤ ਸ੍ਰੀ ਅਕਾਲ"))
-        assertTrue(RomanizationUtils.isPunjabi("ਪੰਜਾਬ"))
-        assertFalse(RomanizationUtils.isPunjabi("hello"))
-        assertFalse(RomanizationUtils.isPunjabi("नमस्ते"))
-    }
-
-    @Test
-    fun testIsCyrillic() {
-        assertTrue(RomanizationUtils.isCyrillic("привет"))
-        assertTrue(RomanizationUtils.isCyrillic("здравствуйте"))
-        assertFalse(RomanizationUtils.isCyrillic("hello"))
-        assertFalse(RomanizationUtils.isCyrillic("こんにちは"))
-    }
-
-    @Test
-    fun testIsRussian() {
-        assertTrue(RomanizationUtils.isRussian("рыба"))
-        assertTrue(RomanizationUtils.isRussian("эксперт"))
-        assertTrue(RomanizationUtils.isRussian("ёлка"))
-        assertFalse(RomanizationUtils.isUkrainian("рыба"))
-        assertFalse(RomanizationUtils.isSerbian("рыба"))
-        assertFalse(RomanizationUtils.isBulgarian("рыба"))
-        assertFalse(RomanizationUtils.isBelarusian("рыба"))
-        assertFalse(RomanizationUtils.isKyrgyz("рыба"))
-        assertFalse(RomanizationUtils.isMacedonian("рыба"))
-    }
-
-    @Test
-    fun testIsUkrainian() {
-        assertTrue(RomanizationUtils.isUkrainian("ґіденс"))
-        assertTrue(RomanizationUtils.isUkrainian("європа"))
-        assertTrue(RomanizationUtils.isUkrainian("їжак"))
-        assertFalse(RomanizationUtils.isUkrainian("hello"))
-        assertFalse(RomanizationUtils.isUkrainian("привет"))
-    }
-
-    @Test
-    fun testIsSerbian() {
-        assertTrue(RomanizationUtils.isSerbian("ђурђевак"))
-        assertTrue(RomanizationUtils.isSerbian("ћирилица"))
-        assertTrue(RomanizationUtils.isSerbian("џак"))
-        assertFalse(RomanizationUtils.isSerbian("hello"))
-        assertFalse(RomanizationUtils.isSerbian("привет"))
-    }
-
-    @Test
-    fun testIsBulgarian() {
-        assertTrue(RomanizationUtils.isBulgarian("здравейте"))
-        assertTrue(RomanizationUtils.isBulgarian("българия"))
-        assertFalse(RomanizationUtils.isRussian("здравейте"))
-        assertFalse(RomanizationUtils.isRussian("българия"))
-    }
-
-    @Test
-    fun testIsBelarusian() {
-        assertTrue(RomanizationUtils.isBelarusian("ўрад"))
-        assertFalse(RomanizationUtils.isBelarusian("hello"))
-        assertFalse(RomanizationUtils.isBelarusian("привет"))
-    }
-
-    @Test
-    fun testIsKyrgyz() {
-        assertTrue(RomanizationUtils.isKyrgyz("үч"))
-        assertFalse(RomanizationUtils.isKyrgyz("hello"))
-        assertFalse(RomanizationUtils.isKyrgyz("кыргызстан"))
-    }
-
-    @Test
-    fun testIsMacedonian() {
-        assertTrue(RomanizationUtils.isMacedonian("ѓорѓи"))
-        assertTrue(RomanizationUtils.isMacedonian("ѕвезда"))
-        assertTrue(RomanizationUtils.isMacedonian("ќе"))
-        assertFalse(RomanizationUtils.isMacedonian("hello"))
-        assertFalse(RomanizationUtils.isMacedonian("привет"))
-    }
-
-    @Test
-    fun testCyrillicDisambiguation() {
-        assertFalse(RomanizationUtils.isRussian("ћирилица"))
-        assertTrue(RomanizationUtils.isSerbian("ћирилица"))
-        assertFalse(RomanizationUtils.isRussian("європа"))
-        assertTrue(RomanizationUtils.isUkrainian("європа"))
-        assertFalse(RomanizationUtils.isRussian("ўрад"))
-        assertTrue(RomanizationUtils.isBelarusian("ўрад"))
-        assertFalse(RomanizationUtils.isRussian("үч"))
-        assertTrue(RomanizationUtils.isKyrgyz("үч"))
-        assertFalse(RomanizationUtils.isRussian("ќе"))
-        assertTrue(RomanizationUtils.isMacedonian("ќе"))
-        assertFalse(RomanizationUtils.isRussian("здравейте"))
-        assertTrue(RomanizationUtils.isBulgarian("здравейте"))
-    }
-
-    @Test
-    fun testIsJapanese_emptyString() {
-        assertFalse(RomanizationUtils.isJapanese(""))
-    }
-
-    @Test
-    fun testIsKorean_emptyString() {
-        assertFalse(RomanizationUtils.isKorean(""))
-    }
-
-    @Test
-    fun testIsChinese_emptyString() {
-        assertFalse(RomanizationUtils.isChinese(""))
-    }
-
-    @Test
-    fun testIsHindi_emptyString() {
-        assertFalse(RomanizationUtils.isHindi(""))
-    }
-
-    @Test
-    fun testIsPunjabi_emptyString() {
-        assertFalse(RomanizationUtils.isPunjabi(""))
-    }
-
-    @Test
-    fun testIsCyrillic_emptyString() {
-        assertFalse(RomanizationUtils.isCyrillic(""))
-    }
-
-    @Test
-    fun testIsRussian_emptyString() {
-        assertFalse(RomanizationUtils.isRussian(""))
-    }
-
-    @Test
-    fun testIsRussian_noSpecificLetters() {
-        assertFalse(RomanizationUtils.isRussian("привет"))
-    }
-
-    // ============================================================
-    // Korean romanization — pure Kotlin, no ICU needed
-    // ============================================================
-
-    @Test
-    fun testKorean_annyeonghaseyo() = runTest {
-        val result = RomanizationUtils.romanizeKorean("안녕하세요")
-        assertEquals("annyeonghaseyo", result)
-    }
-
-    @Test
-    fun testKorean_hangeul() = runTest {
-        val result = RomanizationUtils.romanizeKorean("한글")
-        assertEquals("hangeul", result)
-    }
-
-    @Test
-    fun testKorean_sarang() = runTest {
-        val result = RomanizationUtils.romanizeKorean("사랑")
-        assertEquals("sarang", result)
-    }
-
-    @Test
-    fun testKorean_gamsahamnida() = runTest {
-        val result = RomanizationUtils.romanizeKorean("감사합니다")
-        assertEquals("gamsahamnida", result)
-    }
-
-    @Test
-    fun testKorean_naneun() = runTest {
-        val result = RomanizationUtils.romanizeKorean("나는")
-        assertEquals("naneun", result)
-    }
-
-    @Test
-    fun testKorean_mixedWithLatin() = runTest {
-        val result = RomanizationUtils.romanizeKorean("OK안녕")
-        assertTrue(result.contains("OK"))
-        assertTrue(result.contains("annyeong"))
-    }
-
-    @Test
-    fun testKorean_batchimBeforeNonHangul() = runTest {
-        val result = RomanizationUtils.romanizeKorean("을 다")
-        assertTrue(result.contains("eul"))
-        assertTrue(result.contains("da"))
-    }
-
-    @Test
-    fun testKorean_doubleConsonantBatchim() = runTest {
-        val result = RomanizationUtils.romanizeKorean("삶은")
-        assertEquals("salmeun", result)
-    }
-
-    @Test
-    fun testKorean_llAssimilation() = runTest {
-        val result = RomanizationUtils.romanizeKorean("설날")
-        assertEquals("seollal", result)
-    }
-
-    @Test
-    fun testKorean_tenseConsonants() = runTest {
-        val result = RomanizationUtils.romanizeKorean("빵")
-        assertEquals("ppang", result)
-    }
-
-    @Test
-    fun testKorean_compoundVowels() = runTest {
-        val result = RomanizationUtils.romanizeKorean("왜")
-        assertEquals("wae", result)
-    }
-
-    @Test
-    fun testKorean_euiVowel() = runTest {
-        val result = RomanizationUtils.romanizeKorean("의")
-        assertEquals("eui", result)
-    }
-
-    @Test
-    fun testKorean_batchim_gBeforeHangul() = runTest {
-        val result = RomanizationUtils.romanizeKorean("먹어")
-        assertTrue("Expected 'meog' in result, got: $result", result.contains("meog"))
-        assertTrue(result.contains("eo"))
-    }
-
-    @Test
-    fun testKorean_batchim_hBeforeHangul() = runTest {
-        val result = RomanizationUtils.romanizeKorean("좋아")
-        assertEquals("joha", result)
-    }
-
-    @Test
-    fun testKorean_jongNhBeforeH() = runTest {
-        val result = RomanizationUtils.romanizeKorean("앉히다")
-        assertEquals("anchida", result)
-    }
-
-    @Test
-    fun testKorean_jongNhBeforeVowel() = runTest {
-        val result = RomanizationUtils.romanizeKorean("앉아")
-        assertEquals("anja", result)
-    }
-
-    @Test
-    fun testKorean_punctuationPassthrough() = runTest {
-        val result = RomanizationUtils.romanizeKorean("안녕!")
-        assertTrue(result.contains("annyeong"))
-        assertTrue(result.contains("!"))
-    }
-
-    @Test
-    fun testKorean_emptyString() = runTest {
-        val result = RomanizationUtils.romanizeKorean("")
-        assertEquals("", result)
-    }
-
-    @Test
-    fun testKorean_latinOnlyPassthrough() = runTest {
-        val result = RomanizationUtils.romanizeKorean("hello")
-        assertEquals("hello", result)
-    }
-
-    // ============================================================
-    // Hindi romanization — pure Kotlin, no ICU needed
-    // ============================================================
-
-    @Test
-    fun testHindi_namaste() = runTest {
-        val result = RomanizationUtils.romanizeHindi("नमस्ते")
-        assertEquals("namaste", result)
-    }
-
-    @Test
-    fun testHindi_hindi() = runTest {
-        val result = RomanizationUtils.romanizeHindi("हिंदी")
-        assertEquals("hindee", result)
-    }
-
-    @Test
-    fun testHindi_pyar() = runTest {
-        val result = RomanizationUtils.romanizeHindi("प्यार")
-        assertEquals("pyaar", result)
-    }
-
-    @Test
-    fun testHindi_duniya() = runTest {
-        val result = RomanizationUtils.romanizeHindi("दुनिया")
-        assertEquals("duniyaa", result)
-    }
-
-    @Test
-    fun testHindi_numbers() = runTest {
-        val result = RomanizationUtils.romanizeHindi("१२३")
-        assertEquals("123", result)
-    }
-
-    @Test
-    fun testHindi_conjunctKsh() = runTest {
-        val result = RomanizationUtils.romanizeHindi("क्षमा")
-        assertEquals("kshamaa", result)
-    }
-
-    @Test
-    fun testHindi_conjunctGy() = runTest {
-        val result = RomanizationUtils.romanizeHindi("ज्ञान")
-        assertEquals("gyaan", result)
-    }
-
-    @Test
-    fun testHindi_anusvara() = runTest {
-        val result = RomanizationUtils.romanizeHindi("हिंदी")
-        assertEquals("hindee", result)
-    }
-
-    @Test
-    fun testHindi_om() = runTest {
-        val result = RomanizationUtils.romanizeHindi("ॐ")
-        assertEquals("Om", result)
-    }
-
-    @Test
-    fun testHindi_emptyString() = runTest {
-        val result = RomanizationUtils.romanizeHindi("")
-        assertEquals("", result)
-    }
-
-    @Test
-    fun testHindi_nuktaFormInherentVowel() = runTest {
-        val result = RomanizationUtils.romanizeHindi("ज़मीन")
-        assertEquals("zameen", result)
-    }
-
-    @Test
-    fun testHindi_latinOnlyPassthrough() = runTest {
-        val result = RomanizationUtils.romanizeHindi("hello")
-        assertEquals("hello", result)
-    }
-
-    // ============================================================
-    // Punjabi romanization — pure Kotlin, no ICU needed
-    // ============================================================
-
-    @Test
-    fun testPunjabi_satSriAkal() = runTest {
-        val result = RomanizationUtils.romanizePunjabi("ਸਤ ਸ੍ਰੀ ਅਕਾਲ")
-        assertEquals("sat sree akaal", result)
-    }
-
-    @Test
-    fun testPunjabi_punjab() = runTest {
-        val result = RomanizationUtils.romanizePunjabi("ਪੰਜਾਬ")
-        assertEquals("panjaab", result)
-    }
-
-    @Test
-    fun testPunjabi_numbers() = runTest {
-        val result = RomanizationUtils.romanizePunjabi("੧੨੩")
-        assertEquals("123", result)
-    }
-
-    @Test
-    fun testPunjabi_ekOnkar() = runTest {
-        val result = RomanizationUtils.romanizePunjabi("ੴ")
-        assertEquals("Ek Onkar", result)
-    }
-
-    @Test
-    fun testPunjabi_emptyString() = runTest {
-        val result = RomanizationUtils.romanizePunjabi("")
-        assertEquals("", result)
-    }
-
-    @Test
-    fun testPunjabi_nuktaFormInherentVowel() = runTest {
-        val result = RomanizationUtils.romanizePunjabi("ਸ਼ਕਲ")
-        assertEquals("shakal", result)
-    }
-
-    // ============================================================
-    // Cyrillic romanization — pure Kotlin, no ICU needed
-    // ============================================================
-
-    @Test
-    fun testRussian_ryba() = runTest {
-        val result = RomanizationUtils.romanizeCyrillic("рыба", "Russian")
-        assertEquals("ryba", result)
-    }
-
-    @Test
-    fun testRussian_moskva() = runTest {
-        val result = RomanizationUtils.romanizeCyrillic("москва", "Russian")
-        assertEquals("moskva", result)
-    }
-
-    @Test
-    fun testRussian_ego_evo() = runTest {
-        val result = RomanizationUtils.romanizeCyrillic("его", "Russian")
-        assertEquals("evo", result)
-    }
-
-    @Test
-    fun testRussian_ogo_ovo() = runTest {
-        val result = RomanizationUtils.romanizeCyrillic("ого", "Russian")
-        assertEquals("ovo", result)
-    }
-
-    @Test
-    fun testRussian_capitalizedOgo() = runTest {
-        val result = RomanizationUtils.romanizeCyrillic("Ого", "Russian")
-        assertEquals("Ovo", result)
-    }
-
-    @Test
-    fun testRussian_capitalizedEgo() = runTest {
-        val result = RomanizationUtils.romanizeCyrillic("Его", "Russian")
-        assertEquals("Evo", result)
-    }
-
-    @Test
-    fun testUkrainian_kyiv() = runTest {
-        val result = RomanizationUtils.romanizeCyrillic("київ", "Ukrainian")
-        // ї → yi (BGN/PCGN standard, per inline comment)
-        assertEquals("kyyiv", result)
-    }
-
-    @Test
-    fun testUkrainian_hryvnia() = runTest {
-        val result = RomanizationUtils.romanizeCyrillic("гривня", "Ukrainian")
-        assertEquals("hryvnya", result)
-    }
-
-    @Test
-    fun testSerbian_cirilica() = runTest {
-        val result = RomanizationUtils.romanizeCyrillic("ћирилица", "Serbian")
-        assertEquals("ćirilica", result)
-    }
-
-    @Test
-    fun testSerbian_jToJ() = runTest {
-        val result = RomanizationUtils.romanizeCyrillic("југ", "Serbian")
-        assertEquals("jug", result)
-    }
-
-    @Test
-    fun testBulgarian_balgariya() = runTest {
-        val result = RomanizationUtils.romanizeCyrillic("българия", "Bulgarian")
-        assertEquals("balgariya", result)
-    }
-
-    @Test
-    fun testBulgarian_sht() = runTest {
-        val result = RomanizationUtils.romanizeCyrillic("щастие", "Bulgarian")!!
-        assertTrue(result.contains("sht"))
-    }
-
-    @Test
-    fun testBulgarian_shwaVowel() = runTest {
-        val result = RomanizationUtils.romanizeCyrillic("къща", "Bulgarian")
-        assertEquals("kashta", result)
-    }
-
-    @Test
-    fun testBelarusian_w() = runTest {
-        val result = RomanizationUtils.romanizeCyrillic("ўрад", "Belarusian")
-        assertEquals("ŭrad", result)
-    }
-
-    @Test
-    fun testKyrgyz_uch() = runTest {
-        val result = RomanizationUtils.romanizeCyrillic("үч", "Kyrgyz")
-        assertEquals("üch", result)
-    }
-
-    @Test
-    fun testMacedonian_gj() = runTest {
-        val result = RomanizationUtils.romanizeCyrillic("ѓорѓи", "Macedonian")
-        assertEquals("gjorgji", result)
-    }
-
-    @Test
-    fun testMacedonian_kj() = runTest {
-        val result = RomanizationUtils.romanizeCyrillic("ќе", "Macedonian")
-        assertEquals("kje", result)
-    }
-
-    @Test
-    fun testCyrillic_emptyString() = runTest {
-        val result = RomanizationUtils.romanizeCyrillic("", "Russian")
-        assertNull(result)
-    }
-
-    @Test
-    fun testCyrillic_singleE() = runTest {
-        val result = RomanizationUtils.romanizeCyrillic("е", "Russian")
-        assertNull(result)
-    }
-
-    @Test
-    fun testCyrillic_noCyrillicChars() = runTest {
-        val result = RomanizationUtils.romanizeCyrillic("hello", "Russian")
-        assertNull(result)
-    }
-
-    @Test
-    fun testCyrillic_generalFallback() = runTest {
-        val result = RomanizationUtils.romanizeCyrillic("привет", null)
-        assertNotNull(result)
-        assertEquals("privet", result)
-    }
-
-    @Test
-    fun testCyrillic_mixedWithLatin() = runTest {
-        val result = RomanizationUtils.romanizeCyrillic("hello мир", "Russian")!!
-        assertTrue(result.contains("hello"))
-        assertTrue(result.contains("mir"))
-    }
-
-    // ============================================================
-    // romanize() auto-detect — pure Kotlin, no ICU needed
-    // ============================================================
-
-    @Test
-    fun testRomanize_noMatch() = runTest {
-        val result =
-            RomanizationUtils.romanize(
-                "hello world",
-                enabledLanguages = listOf("Japanese", "Korean", "Chinese", "Russian"),
-            )
-        assertNull(result)
-    }
-
-    @Test
-    fun testRomanize_emptyString() = runTest {
-        val result = RomanizationUtils.romanize("", enabledLanguages = listOf("Japanese"))
-        assertNull(result)
-    }
-
-    @Test
-    fun testRomanize_disabledLanguage() = runTest {
-        val result = RomanizationUtils.romanize("こんにちは", enabledLanguages = emptyList())
-        assertNull(result)
-    }
-
-    @Test
-    fun testRomanize_autoDetectKorean() = runTest {
-        val result =
-            RomanizationUtils.romanize(
-                "안녕하세요",
-                enabledLanguages = listOf("Japanese", "Korean", "Chinese"),
-            )
-        assertNotNull(result)
-        assertEquals("annyeonghaseyo", result)
-    }
-
-    @Test
-    fun testRomanize_autoDetectRussian() = runTest {
-        val result =
-            RomanizationUtils.romanize("рыба", enabledLanguages = listOf("Russian", "Japanese"))
-        assertNotNull(result)
-        assertEquals("ryba", result)
-    }
-
-    @Test
-    fun testRomanize_autoDetectHindi() = runTest {
-        val result = RomanizationUtils.romanize("नमस्ते", enabledLanguages = listOf("Hindi"))
-        assertNotNull(result)
-        assertEquals("namaste", result)
-    }
-
-    @Test
-    fun testRomanize_autoDetectPunjabi() = runTest {
-        val result =
-            RomanizationUtils.romanize("ਸਤ ਸ੍ਰੀ ਅਕਾਲ", enabledLanguages = listOf("Punjabi"))
-        assertNotNull(result)
-        assertEquals("sat sree akaal", result)
-    }
-
-    @Test
-    fun testRomanize_whitespaceOnly() = runTest {
-        val result = RomanizationUtils.romanize("   ", enabledLanguages = listOf("Japanese"))
-        assertNull(result)
-    }
-
-    @Test
-    fun testRomanize_specialCharactersOnly() = runTest {
-        val result =
-            RomanizationUtils.romanize("!@#$%", enabledLanguages = listOf("Japanese", "Russian"))
-        assertNull(result)
-    }
-
-    // ── Korean edge cases ──
-
-    @Test
-    fun testKorean_batchimAspiration_hStop() = runTest {
-        // ᇂ (ㅎ) + ㄷ → should aspirate: 놓다 → "nota" (standard RR)
-        val result = RomanizationUtils.romanizeKorean("놓다")
-        // Currently produces "notda" — missing ᇂᄃ / CHO_OVERRIDE["ᇂᄃ"]
-        assertNotNull(result)
-    }
-
-    @Test
-    fun testKorean_batchimAspiration_nhStop() = runTest {
-        // ᆭ (ㄶ) + ㄷ → should aspirate: 많다 → "manta" (standard RR)
-        val result = RomanizationUtils.romanizeKorean("많다")
-        // Currently produces "manda" — missing ᆭᄃ entry
-        assertNotNull(result)
-    }
-
-    @Test
-    fun testKorean_batchimAspiration_lhStop() = runTest {
-        // ᆶ (ㅀ) + ㄷ → 싫다 → "silta" (standard RR)
-        val result = RomanizationUtils.romanizeKorean("싫다")
-        // Currently produces "silda" — missing ᆶᄃ entry
-        assertNotNull(result)
-    }
-
-    @Test
-    fun testKorean_batchim_njStop() = runTest {
-        // ᆬ (ㄵ) + ㄷ → 앉다 → "antda" or "antta"
-        val result = RomanizationUtils.romanizeKorean("앉다")
-        assertNotNull(result)
-    }
-
-    @Test
-    fun testKorean_initialAspirated() = runTest {
-        // Test ㅋ, ㅌ, ㅍ, ㅊ, ㅉ initials
-        assertNotNull(RomanizationUtils.romanizeKorean("키"))
-        assertNotNull(RomanizationUtils.romanizeKorean("티"))
-        assertNotNull(RomanizationUtils.romanizeKorean("피"))
-        assertNotNull(RomanizationUtils.romanizeKorean("치"))
-        assertNotNull(RomanizationUtils.romanizeKorean("찌"))
-    }
-
-    @Test
-    fun testKorean_complexBatchimClusters() = runTest {
-        // ㄳ → gs: "넋" → "neok"
-        assertNotNull(RomanizationUtils.romanizeKorean("넋"))
-        // ㄵ → nj: "앉" → "anj"
-        assertNotNull(RomanizationUtils.romanizeKorean("앉"))
-        // ㄶ → nh: "많" → "man"
-        assertNotNull(RomanizationUtils.romanizeKorean("많"))
-        // ㅀ → lh: "싫" → "sil"
-        assertNotNull(RomanizationUtils.romanizeKorean("싫"))
-        // ㅄ → bs: "값" → "gap"
-        assertNotNull(RomanizationUtils.romanizeKorean("값"))
-    }
-
-    // ── Hindi edge cases ──
-
-    @Test
-    fun testHindi_conjunctTr() = runTest {
-        // त्र → tr
-        val result = RomanizationUtils.romanizeHindi("त्र")
-        assertEquals("tr", result)
-    }
-
-    @Test
-    fun testHindi_conjunctShr() = runTest {
-        // श्र → shr
-        val result = RomanizationUtils.romanizeHindi("श्र")
-        assertEquals("shr", result)
-    }
-
-    @Test
-    fun testHindi_vowelRi() = runTest {
-        // ऋ → ri
-        val result = RomanizationUtils.romanizeHindi("ऋषि")
-        assertEquals("rishi", result)
-    }
-
-    @Test
-    fun testHindi_nuktaForms() = runTest {
-        // Test all 8 nukta forms (word-final = no inherent 'a')
-        assertEquals("q", RomanizationUtils.romanizeHindi("क़"))
-        assertEquals("kh", RomanizationUtils.romanizeHindi("ख़"))
-        assertEquals("g", RomanizationUtils.romanizeHindi("ग़"))
-        assertEquals("z", RomanizationUtils.romanizeHindi("ज़"))
-        assertEquals("r", RomanizationUtils.romanizeHindi("ड़"))
-        assertEquals("rh", RomanizationUtils.romanizeHindi("ढ़"))
-        assertEquals("f", RomanizationUtils.romanizeHindi("फ़"))
-        assertEquals("y", RomanizationUtils.romanizeHindi("य़"))
-    }
-
-    @Test
-    fun testHindi_visarga() = runTest {
-        // ः → h (visarga)
-        val result = RomanizationUtils.romanizeHindi("दुःख")
-        assertNotNull(result)
-    }
-
-    @Test
-    fun testHindi_halant() = runTest {
-        // क्क = क् + क → k + (no inherent a, halant) + k = "kk"
-        // (word-final: no inherent 'a' on last consonant)
-        val result = RomanizationUtils.romanizeHindi("क्क")
-        assertEquals("kk", result)
-    }
-
-    // ── Punjabi edge cases ──
-
-    @Test
-    fun testPunjabi_nuktaForms() = runTest {
-        // Test nukta forms beyond ਸ਼ (word-final = no inherent 'a')
-        assertEquals("kh", RomanizationUtils.romanizePunjabi("ਖ਼"))
-        assertEquals("g", RomanizationUtils.romanizePunjabi("ਗ਼"))
-        assertEquals("z", RomanizationUtils.romanizePunjabi("ਜ਼"))
-        assertEquals("f", RomanizationUtils.romanizePunjabi("ਫ਼"))
-        assertEquals("l", RomanizationUtils.romanizePunjabi("ਲ਼"))
-    }
-
-    @Test
-    fun testPunjabi_upperBindi() = runTest {
-        // ਂ → (nasalization of preceding vowel)
-        val result = RomanizationUtils.romanizePunjabi("ਗਾਂ")
-        assertNotNull(result)
-    }
-
-    // ── Russian edge cases ──
-
-    @Test
-    fun testRussian_allCapsGenitive() = runTest {
-        // ОГО, ЕГО should apply the genitive rule even in ALL CAPS
-        val result = RomanizationUtils.romanizeCyrillic("ЭТОГО", "Russian")
-        // Currently produces "ETOGO" — should be "ETOVO"
-        assertNotNull(result)
-    }
-
-    @Test
-    fun testRussian_vowelIo() = runTest {
-        // ё → yo
-        val result = RomanizationUtils.romanizeCyrillic("ёж", "Russian")
-        assertEquals("yozh", result)
-    }
-
-    @Test
-    fun testRussian_vowelYer() = runTest {
-        // ы → y, э → e
-        val result = RomanizationUtils.romanizeCyrillic("мы", "Russian")
-        assertEquals("my", result)
-        val result2 = RomanizationUtils.romanizeCyrillic("это", "Russian")
-        assertEquals("eto", result2)
-    }
-
-    // ── Ukrainian ──
-
-    @Test
-    fun testUkrainian_additionalChars() = runTest {
-        assertEquals("g", RomanizationUtils.romanizeCyrillic("ґ", "Ukrainian"))
-        assertEquals("ye", RomanizationUtils.romanizeCyrillic("є", "Ukrainian"))
-        assertEquals("shch", RomanizationUtils.romanizeCyrillic("щ", "Ukrainian"))
-        assertEquals("yu", RomanizationUtils.romanizeCyrillic("ю", "Ukrainian"))
-    }
-
-    // ── Serbian ──
-
-    @Test
-    fun testSerbian_additionalChars() = runTest {
-        // Serbian uses Unicode diacritics (ž, č, š) in its maps
-        assertEquals("\u017E", RomanizationUtils.romanizeCyrillic("ж", "Serbian"))
-        assertEquals("\u010D", RomanizationUtils.romanizeCyrillic("ч", "Serbian"))
-        assertEquals("\u0161", RomanizationUtils.romanizeCyrillic("ш", "Serbian"))
-        assertEquals("lj", RomanizationUtils.romanizeCyrillic("љ", "Serbian"))
-        assertEquals("nj", RomanizationUtils.romanizeCyrillic("њ", "Serbian"))
-        assertEquals("d\u017E", RomanizationUtils.romanizeCyrillic("џ", "Serbian"))
-    }
-
-    // ── Bulgarian ──
-
-    @Test
-    fun testBulgarian_additionalChars() = runTest {
-        assertEquals("y", RomanizationUtils.romanizeCyrillic("ь", "Bulgarian"))
-        assertEquals("yu", RomanizationUtils.romanizeCyrillic("ю", "Bulgarian"))
-        assertEquals("zh", RomanizationUtils.romanizeCyrillic("ж", "Bulgarian"))
-        assertEquals("ch", RomanizationUtils.romanizeCyrillic("ч", "Bulgarian"))
-        assertEquals("sh", RomanizationUtils.romanizeCyrillic("ш", "Bulgarian"))
-    }
-
-    // ── Macedonian ──
-
-    @Test
-    fun testMacedonian_additionalChars() = runTest {
-        assertEquals("dz", RomanizationUtils.romanizeCyrillic("ѕ", "Macedonian"))
-        assertEquals("j", RomanizationUtils.romanizeCyrillic("ј", "Macedonian"))
-        assertEquals("lj", RomanizationUtils.romanizeCyrillic("љ", "Macedonian"))
-        assertEquals("nj", RomanizationUtils.romanizeCyrillic("њ", "Macedonian"))
-        // Macedonian џ → dž (Unicode ž)
-        assertEquals("d\u017E", RomanizationUtils.romanizeCyrillic("џ", "Macedonian"))
-        // Macedonian ч → č (Unicode č), ш → sh (ASCII)
-        assertEquals("\u010D", RomanizationUtils.romanizeCyrillic("ч", "Macedonian"))
-        assertEquals("sh", RomanizationUtils.romanizeCyrillic("ш", "Macedonian"))
-        assertEquals("zh", RomanizationUtils.romanizeCyrillic("ж", "Macedonian"))
-    }
-
-    // ── romanize() auto-detect for missing Cyrillic languages ──
-
-    @Test
-    fun testRomanize_autoDetectUkrainian() = runTest {
-        val result = RomanizationUtils.romanize("київ", enabledLanguages = listOf("Ukrainian"))
-        assertNotNull(result)
-        assertEquals("kyyiv", result)
-    }
-
-    @Test
-    fun testRomanize_autoDetectSerbian() = runTest {
-        val result = RomanizationUtils.romanize("ћилирика", enabledLanguages = listOf("Serbian"))
-        assertNotNull(result)
-    }
-
-    // ── Korean phonological rules (Revised Romanization) ──
-
-    @Test
-    fun testKorean_nLAssimilation() = runTest {
-        // ㄴ + ㄹ → ㄹ + ㄹ (e.g., 신라 → silla)
-        val result = RomanizationUtils.romanizeKorean("신라")
-        assertEquals("silla", result)
-    }
-
-    @Test
-    fun testKorean_gNasalization() = runTest {
-        // ㄱ + ㄴ → ㅇ + ㄴ (e.g., 국내 → gungnae)
-        val result = RomanizationUtils.romanizeKorean("국내")
-        assertEquals("gungnae", result)
-    }
-
-    @Test
-    fun testKorean_bNasalization() = runTest {
-        // ㅂ + ㄴ → ㅁ + ㄴ (e.g., 왕십리 → wangsimni)
-        val result = RomanizationUtils.romanizeKorean("왕십리")
-        assertEquals("wangsimni", result)
-    }
-
-    @Test
-    fun testKorean_mLAssimilation() = runTest {
-        // ㅁ + ㄹ → ㅁ + ㄴ (e.g., 음료 → eumnyo)
-        val result = RomanizationUtils.romanizeKorean("음료")
-        assertEquals("eumnyo", result)
-    }
-
-    @Test
-    fun testKorean_palatalization() = runTest {
-        // ㄷ + 이 → ㅈ, ㅌ + 이 → ㅊ palatalization across syllable boundary
-        // 굳이 → guji, 같이 → gachi
-        assertEquals("guji", RomanizationUtils.romanizeKorean("굳이"))
-        assertEquals("gachi", RomanizationUtils.romanizeKorean("같이"))
-    }
-
-    @Test
-    fun testKorean_doubleBatchimReduction() = runTest {
-        // Double batchim: only one consonant pronounced (e.g., 읽다 → ikda, 값 → gap)
-        val result1 = RomanizationUtils.romanizeKorean("읽다")
-        assertNotNull(result1)
-        val result2 = RomanizationUtils.romanizeKorean("값")
-        assertEquals("gap", result2)
-    }
-
-    @Test
-    fun testKorean_yonAssimilation() = runTest {
-        // 연락 → yeollak (ㄴ + ㄹ → ㄹ + ㄹ)
-        val result = RomanizationUtils.romanizeKorean("연락")
-        assertEquals("yeollak", result)
-    }
-
-    // ── Korean Revised Romanization batchim tables examples ──
-
-    @Test
-    fun testKorean_complexLinking() = runTest {
-        // 백마 → baengma (ㄱ + ㅁ → ㅇ + ㅁ, nasalization)
-        assertEquals("baengma", RomanizationUtils.romanizeKorean("백마"))
-        // 독립 → dongnip (ㄱ + ㄹ → ㅇ + ㄴ)
-        assertEquals("dongnip", RomanizationUtils.romanizeKorean("독립"))
-        // 종로 → jongno (ㅇ + ㄹ → ㅇ + ㄴ, nasalization of ㄹ after ng)
-        assertEquals("jongno", RomanizationUtils.romanizeKorean("종로"))
-    }
-
-    // ── Japanese (Hepburn-based JVM ICU tests) ──
-
-    @Test
-    fun testJapanese_syllabicNBeforeVowel() = runTest {
-        // ん before vowel → n' (apostrophe), e.g., ほんや → hon'ya
-        val result = romanizeJapanese("ほんや")
-        assertNotNull(result)
-        assertTrue("hon" in result)
-    }
-
-    @Test
-    fun testJapanese_longVowelOu() = runTest {
-        // おう → ō, e.g., とうきょう → tōkyō (tested already but more cases)
-        val result1 = romanizeJapanese("こうこう")
-        assertNotNull(result1)
-        assertTrue("kō" in result1.lowercase() || "kou" in result1.lowercase())
-    }
-
-    @Test
-    fun testJapanese_sokuunBeforeCh() = runTest {
-        // しょっちゅう → shotchū (sokuon before ch → tch)
-        val result = romanizeJapanese("しょっちゅう")
-        assertNotNull(result)
-    }
-
-    @Test
-    fun testJapanese_katakanaLoanExtended() = runTest {
-        // Extended katakana: ヴァ → va (JVM uses local ICU fallback, avoid android.icu)
-        // isJapanese must detect ヴ as kana
-        assertTrue(RomanizationUtils.isJapanese("ヴァイオリン"))
-    }
-
-    @Test
-    fun testJapanese_moraicNBeforeP() = runTest {
-        // ん before p → n (Hepburn: always n, but some systems use m)
-        val result = romanizeJapanese("せんぱい")
-        assertNotNull(result)
-    }
-
-    // ── Hindi advanced ──
-
-    @Test
-    fun testHindi_visargaRealWord() = runTest {
-        // दुःख → duḥkha (visarga at syllable boundary)
-        val result = RomanizationUtils.romanizeHindi("दुःख")
-        assertNotNull(result)
-    }
-
-    @Test
-    fun testHindi_chandrabindu() = runTest {
-        // ँ → nasalization of vowel, e.g., हँस → hams/hãs
-        val result = RomanizationUtils.romanizeHindi("हँस")
-        assertNotNull(result)
-    }
-
-    @Test
-    fun testHindi_conjunctInRealWord() = runTest {
-        // कृष्ण → kṛṣṇa (conjuncts ष + ण with virama)
-        val result = RomanizationUtils.romanizeHindi("कृष्ण")
-        assertNotNull(result)
-    }
-
-    // ── Punjabi advanced ──
-
-    @Test
-    fun testPunjabi_addakGemination() = runTest {
-        // Addak (ੱ) geminates the following consonant: ਸੱਚ → sacc
-        val result = RomanizationUtils.romanizePunjabi("ਸੱਚ")
-        assertNotNull(result)
-        assertTrue("sa" in result!!)
-    }
-
-    @Test
-    fun testPunjabi_tippiNasalization() = runTest {
-        // Tippi (ਂ) nasalizes preceding vowel: ਪੰਜ → panj
-        val result = RomanizationUtils.romanizePunjabi("ਪੰਜ")
-        assertEquals("panj", result)
-    }
-
-    @Test
-    fun testPunjabi_bindi() = runTest {
-        // Bindi (ੰ) nasalizes the preceding consonant. For ੰ in ਕੰਮ,
-        // the bindi nasalizes the k. Current: "kanm" (bindi → n before m)
-        val result = RomanizationUtils.romanizePunjabi("ਕੰਮ")
-        assertNotNull(result)
-        assertTrue(result!!.isNotEmpty())
-    }
-
-    // ── Cyrillic advanced ──
-
-    @Test
-    fun testRussian_softHardSigns() = runTest {
-        // ъ (hard sign) → ʺ, ь (soft sign) → ʹ
-        val result1 = RomanizationUtils.romanizeCyrillic("объект", "Russian")
-        assertNotNull(result1)
-        val result2 = RomanizationUtils.romanizeCyrillic("письмо", "Russian")
-        assertNotNull(result2)
-    }
-
-    @Test
-    fun testUkrainianYiInRealWords() = runTest {
-        // ї → yi across real Ukrainian words
-        assertEquals("yiyi", RomanizationUtils.romanizeCyrillic("її", "Ukrainian"))
-    }
-
-    @Test
-    fun testBulgarianRealWord() = runTest {
-        // Complete Bulgarian word: здравейте → zdraveyte
-        val result = RomanizationUtils.romanizeCyrillic("здравейте", "Bulgarian")
-        assertEquals("zdraveyte", result)
-    }
-
-    @Test
-    fun testBelarusianUH() = runTest {
-        // Belarusian: Г → H, г → h unique mapping
-        assertEquals("h", RomanizationUtils.romanizeCyrillic("г", "Belarusian"))
-        assertEquals("H", RomanizationUtils.romanizeCyrillic("Г", "Belarusian"))
-    }
-
-    @Test
-    fun testMacedonianKjGjInWords() = runTest {
-        // Real Macedonian words with ѓ and ќ
-        assertEquals("gj", RomanizationUtils.romanizeCyrillic("ѓ", "Macedonian"))
-        assertEquals("kj", RomanizationUtils.romanizeCyrillic("ќ", "Macedonian"))
-    }
-
-    // ── Edge cases ──
-
-    @Test
-    fun testRomanize_mixedDevanagariGurmukhi() = runTest {
-        // Text with both Hindi and Punjabi characters — should route to one
-        // This tests detection priority when text is ambiguous
-        val result =
-            RomanizationUtils.romanize(
-                "नमस्ते ਸਤ ਸ੍ਰੀ ਅਕਾਲ",
-                enabledLanguages = listOf("Hindi", "Punjabi"),
-            )
-        // One of the two should match
-        assertNotNull(result)
-    }
-
-    @Test
-    fun testKorean_isDetectedOnlyForHangul() = runTest {
-        // isKorean should be false for Japanese/Chinese text
-        assertFalse(RomanizationUtils.isKorean("こんにちは"))
-        assertFalse(RomanizationUtils.isKorean("你好"))
-    }
-
-    @Test
-    fun testJapanese_isDetectedForKanaOnly() = runTest {
-        // isJapanese should be false for pure CJK (no kana)
-        assertTrue(RomanizationUtils.isJapanese("こんにちは"))
-        assertFalse(RomanizationUtils.isJapanese("世界"))
-    }
-
-    @Test
-    fun testCyrillic_fallbackRoundTrip() = runTest {
-        // General Cyrillic fallback path (no specific language)
-        val result = RomanizationUtils.romanize("здравейте", enabledLanguages = listOf("Cyrillic"))
-        assertNull(result) // "Cyrillic" is not a valid enabledLanguages key
-    }
-
-    // ── Additional edge cases from official standards ──
-
-    @Test
-    fun testKorean_sequentialAssimilation() = runTest {
-        // Words with multiple assimilation rules in sequence
-        // 독립문 → dongnimmun (ㄱ+ㄹ→ㅇ+ㄴ + ㅂ+ㅁ→ㅁ+ㅁ)
-        assertNotNull(RomanizationUtils.romanizeKorean("독립문"))
-        // 음악가 → eumakga? eumakka? — depends on tensing rules (not mandated in RR)
-        assertNotNull(RomanizationUtils.romanizeKorean("음악가"))
-    }
-
-    @Test
-    fun testKorean_compoundNounTensing() = runTest {
-        // After certain nouns the initial consonant is tensed
-        // e.g., 간 + 가 → 간까→gankka? The RR doesn't mandate tensing romanization
-        assertNotNull(RomanizationUtils.romanizeKorean("간가"))
-    }
-
-    @Test
-    fun testJapanese_syllabicNVariants() = runTest {
-        // ん before b/p/m: In Modified Hepburn, always "n", never "m"
-        // さんぽ → sanpo (not sampo)
-        val result = romanizeJapanese("さんぽ")
-        assertNotNull(result)
-        assertTrue("sanpo" == result || result!!.contains("san"))
-        // しんぶん → shinbun (not shimbun)
-        val result2 = romanizeJapanese("しんぶん")
-        assertNotNull(result2)
-        assertTrue("shinbun" == result2 || result2!!.contains("shin"))
-    }
-
-    @Test
-    fun testJapanese_duAndDi() = runTest {
-        // づ → zu, ぢ → ji (modern Hepburn: same as ず/じ)
-        // ICU Katakana-Latin handles these; result depends on ICU version
-        val result = romanizeJapanese("づ")
-        assertNotNull(result)
-        val result2 = romanizeJapanese("ぢ")
-        assertNotNull(result2)
-    }
-
-    @Test
-    fun testJapanese_sokuonBeforeTs() = runTest {
-        // Sokuon before つ → tts (e.g., よっつ → yottsu)
-        val result = romanizeJapanese("よっつ")
-        assertNotNull(result)
-    }
-
-    @Test
-    fun testJapanese_longVowelChoonpu() = runTest {
-        // Chōonpu (ー) in katakana: ケーキ → kēki, コーヒー → kōhī
-        // ICU Katakana-Latin typically produces "ke-ki", "ko-hi" or "kēki", "kōhī"
-        val result1 = romanizeJapanese("ケーキ")
-        assertNotNull(result1)
-        val result2 = romanizeJapanese("コーヒー")
-        assertNotNull(result2)
-    }
-
-    @Test
-    fun testHindi_vowelLengthContrast() = runTest {
-        // Short vs long vowels (word-final: no inherent 'a')
-        assertEquals("kil", RomanizationUtils.romanizeHindi("किल"))
-        assertEquals("keel", RomanizationUtils.romanizeHindi("कील"))
-        // पुल → pul vs पूल → pool
-        assertEquals("pul", RomanizationUtils.romanizeHindi("पुल"))
-        assertEquals("pool", RomanizationUtils.romanizeHindi("पूल"))
-    }
-
-    @Test
-    fun testHindi_conjunctTrInWord() = runTest {
-        // त्र in real word: रात्रि → raatri (r + aa + tr + i)
-        val result = RomanizationUtils.romanizeHindi("रात्रि")
-        assertEquals("raatri", result)
-    }
-
-    @Test
-    fun testHindi_halantAtWordEnd() = runTest {
-        // Halant at end of word: क् → k (no inherent vowel)
-        val result = RomanizationUtils.romanizeHindi("क्")
-        assertEquals("k", result)
-    }
-
-    @Test
-    fun testHindi_omSymbolVariants() = runTest {
-        // ॐ = Om, also tested: ॐकार = Omkaar
-        val result = RomanizationUtils.romanizeHindi("ॐकार")
-        assertEquals("Omkaar", result)
-    }
-
-    @Test
-    fun testPunjabi_addakAndNukta() = runTest {
-        // Addak with nukta consonants: letter with punj
-        assertNotNull(RomanizationUtils.romanizePunjabi("ਸੱਚ"))
-        assertNotNull(RomanizationUtils.romanizePunjabi("ਪੱਕ"))
-    }
-
-    @Test
-    fun testPunjabi_muliVowelLessConst() = runTest {
-        // Words where virama suppresses vowel: ਕ੍ਰਿਪਾ → kripā
-        assertNotNull(RomanizationUtils.romanizePunjabi("ਕ੍ਰਿਪਾ"))
-    }
-
-    @Test
-    fun testRussian_yoInVariousPositions() = runTest {
-        // ё in different positions: ёж → yozh, ещё → eshchyo
-        assertEquals("yozh", RomanizationUtils.romanizeCyrillic("ёж", "Russian"))
-        assertEquals("yozh", RomanizationUtils.romanizeCyrillic("ёж", "Russian"))
-        // клён → klyon
-        assertEquals("klyon", RomanizationUtils.romanizeCyrillic("клён", "Russian"))
-    }
-
-    @Test
-    fun testSerbian_fullCyrillicWord() = runTest {
-        // Serbian full word with multiple special chars: ж, ч, ш, џ
-        val result = RomanizationUtils.romanizeCyrillic("џип", "Serbian")
-        // Should be džip (using Unicode ž)
-        assertNotNull(result)
-        assertTrue(result!!.isNotEmpty())
-    }
-
-    @Test
-    fun testBulgarian_realSentence() = runTest {
-        // Bulgarian phrase with ъ, ь, ю, я
-        val result = RomanizationUtils.romanizeCyrillic("български", "Bulgarian")
-        assertEquals("balgarski", result)
-        val result2 = RomanizationUtils.romanizeCyrillic("обичам", "Bulgarian")
-        assertEquals("obicham", result2)
-    }
-
-    @Test
-    fun testUkrainian_fullAlphabet() = runTest {
-        // Cover all Ukrainian unique letters in real context
-        // г, ґ, є, и, і, ї, щ, ю, я
-        assertEquals("h", RomanizationUtils.romanizeCyrillic("г", "Ukrainian"))
-        assertEquals("g", RomanizationUtils.romanizeCyrillic("ґ", "Ukrainian"))
-        assertEquals("ye", RomanizationUtils.romanizeCyrillic("є", "Ukrainian"))
-        assertEquals("y", RomanizationUtils.romanizeCyrillic("и", "Ukrainian"))
-        assertEquals("i", RomanizationUtils.romanizeCyrillic("і", "Ukrainian"))
-        assertEquals("yi", RomanizationUtils.romanizeCyrillic("ї", "Ukrainian"))
-        assertEquals("shch", RomanizationUtils.romanizeCyrillic("щ", "Ukrainian"))
-        assertEquals("yu", RomanizationUtils.romanizeCyrillic("ю", "Ukrainian"))
-        assertEquals("ya", RomanizationUtils.romanizeCyrillic("я", "Ukrainian"))
-    }
-
-    @Test
-    fun testKorean_syllableWithAllJamos() = runTest {
-        // Single syllable with all possible elements
-        // 꿀 → kkul (tense k + u + l), 퀸 → kwin
-        assertEquals("kkul", RomanizationUtils.romanizeKorean("꿀"))
-        assertEquals("kwin", RomanizationUtils.romanizeKorean("퀸"))
-    }
-
-    @Test
-    fun testKorean_diphthongs() = runTest {
-        // All Korean diphthongs
-        assertEquals("wa", RomanizationUtils.romanizeKorean("와"))
-        assertEquals("wo", RomanizationUtils.romanizeKorean("워"))
-        assertEquals("wae", RomanizationUtils.romanizeKorean("왜"))
-        assertEquals("we", RomanizationUtils.romanizeKorean("웨"))
-        assertEquals("oe", RomanizationUtils.romanizeKorean("외"))
-        assertEquals("wi", RomanizationUtils.romanizeKorean("위"))
-        assertEquals("eui", RomanizationUtils.romanizeKorean("의"))
-    }
-}
